refactor(registry): split supply chain conformance
